python script to obtain contrib list of GH repo
This file is a Python script that can be used to obtain a file with the login names of all contributors to a given GitHub repository. It just requires as input the repo owner and repo name, and the name of the output file to be just. Based on a call to the GitHub API it collects all the login names of all contributors of the repo provided as input. I used this script for my own experiments with RABBIT to test it on several GitHub repositories.